+++ src/circle.c
@@ -3,7 +3,7 @@
 /*
  *
  */
-Circle *createCircle(float A[2], float B[2], float C[2])
+Circle *createCircle(double A[2], double B[2], double C[2])
 {
   if (fabs(A[0] - B[0]) < EPSILON && fabs(A[1] - B[1]) < EPSILON)
   {
@@ -24,12 +24,12 @@ Circle *createCircle(float A[2], float B[2], float C[2])
 
   // http://www.ambrsoft.com/trigocalc/circle3d.htm
   // Todo store same computation to be faster and check
-  float p1 = A[0] * (B[1] - C[1]) - A[1] * (B[0] - C[0]) + B[0] * C[1] - B[1] * C[0];
-  float p2 = (A[0] * A[0] + A[1] * A[1]) * (C[1] - B[1]) + (B[0] * B[0] + B[1] * B[1]) * (A[1] - C[1]) + (C[0] * C[0] + C[1] * C[1]) * (B[1] - A[1]);
-  float p3 = (A[0] * A[0] + A[1] * A[1]) * (B[0] - C[0]) + (B[0] * B[0] + B[1] * B[1]) * (C[0] - A[0]) + (C[0] * C[0] + C[1] * C[1]) * (A[0] - B[0]);
-  float p4 = (A[0] * A[0] + A[1] * A[1]) * (C[0] * B[1] - B[0] * C[1]) + (B[0] * B[0] + B[1] * B[1]) * (A[0] * C[1] - C[0] * A[1]) + (C[0] * C[0] + C[1] * C[1]) * (B[0] * A[1] - A[0] * B[1]);
+  double p1 = A[0] * (B[1] - C[1]) - A[1] * (B[0] - C[0]) + B[0] * C[1] - B[1] * C[0];
+  double p2 = (A[0] * A[0] + A[1] * A[1]) * (C[1] - B[1]) + (B[0] * B[0] + B[1] * B[1]) * (A[1] - C[1]) + (C[0] * C[0] + C[1] * C[1]) * (B[1] - A[1]);
+  double p3 = (A[0] * A[0] + A[1] * A[1]) * (B[0] - C[0]) + (B[0] * B[0] + B[1] * B[1]) * (C[0] - A[0]) + (C[0] * C[0] + C[1] * C[1]) * (A[0] - B[0]);
+  double p4 = (A[0] * A[0] + A[1] * A[1]) * (C[0] * B[1] - B[0] * C[1]) + (B[0] * B[0] + B[1] * B[1]) * (A[0] * C[1] - C[0] * A[1]) + (C[0] * C[0] + C[1] * C[1]) * (B[0] * A[1] - A[0] * B[1]);
 
-  float Delta = p2 * p2 + p3 * p3 - 4 * p1 * p4;
+  double Delta = p2 * p2 + p3 * p3 - 4 * p1 * p4;
 
   if (fabs(p1) < EPSILON)
   {
